services/gateway_service/requirements.txt
Django
djangorestframework
graphene-django
httpx
django-cors-headers
PyJWT
 